# api-spec.yml
openapi: 3.0.3
info:
  title: Luminaire Solar API
  description: Provides access to the Luminaire Solar API
  version: '1.0'
components:
  securitySchemes:
    BearerAuth:
      description: RSA256 JWT signed by secret key, with user in payload
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    summary:
      type: object
      description:
        Aggregated energy metrics summary for a specific date, providing
        totals of energy production and consumption. This schema is used for
        reporting and analysis of energy patterns over time.
      properties:
        date:
          type: string
          format: date
          description: The specific date for which this energy summary applies
        total_energy_produced:
          type: number
          description: Total energy produced by the solar system on this date in
            kilowatt-hours (kWh)
        total_energy_consumed:
          type: number
          description: Total energy consumed by the property on this date in
            kilowatt-hours (kWh)
      required:
        - total_energy_produced
        - total_energy_consumed
    allSummary:
      type: object
      description:
        Comprehensive collection of energy summaries across different time
        periods (daily, past week, past month). This schema provides a complete
        overview of energy performance for trend analysis and reporting.
      properties:
        daily:
          type: object
          $ref: '#/components/schemas/summary'
          description: Collection of daily energy summaries showing day-by-day performance
        weekly:
          type: object
          $ref: '#/components/schemas/summary'
          description: Weekly energy summary showing week-by-week performance
        monthly:
          type: object
          $ref: '#/components/schemas/summary'
          description: Monthly energy summary showing month-by-month performance
      required:
        - daily
        - weekly
        - monthly
    forecast:
      type: object
      properties:
        date:
          type: string
          format: date
        irradiation:
          type: number
    error:
      type: object
      description:
        Standardized error response format for API error handling. This
        schema provides consistent error information including status codes and
        messages for troubleshooting and user feedback.
      properties:
        statusCode:
          type: number
          description: HTTP status code indicating the type of error encountered
        error:
          type: string
          description: Error type or category for classification purposes
        message:
          type: string
          description: Human-readable error message explaining what went wrong
      required:
        - statusCode
        - error
        - message
...
paths:
...
  /api/summary/{systemId}:
    get:
      operationId: getMetricsSummaryBySystem
      tags:
        - metrics
      description:
        Retrieves aggregated energy summaries for a solar system across
        multiple time periods (daily, past week, past month). This endpoint
        provides a comprehensive overview of energy production and consumption
        patterns, enabling trend analysis and performance evaluation over
        different timeframes.
      parameters:
        - schema:
            type: string
            format: date
          in: query
          name: date
          required: false
          description:
            Reference date (YYYY-MM-DD format) for generating summaries. If not
            provided, defaults to current date.
        - schema:
            type: string
          in: path
          name: systemId
          required: true
          description:
            Unique identifier (UUID) of the solar system to retrieve summary
            data for
      security:
        - BearerAuth: []
      responses:
        '200':
          description:
            Successfully retrieved energy summaries for the specified solar
            system. The response includes aggregated data organized into daily,
            weekly, and monthly collections, each containing total energy
            production and consumption values.
          content:
            application/json:
              schema:
                description:
                  Successfully retrieved energy summaries for the specified solar
                  system. The response includes aggregated data organized into
                  daily, weekly, and monthly collections, each containing total
                  energy production and consumption values.
                type: object
                $ref: '#/components/schemas/allSummary'
        '500':
          description:
            Server encountered an unexpected error while generating summary
            data. This may occur due to database connectivity issues,
            insufficient metric data, or calculation errors.
          content:
            application/json:
              schema:
                description:
                  Server encountered an unexpected error while generating summary
                  data. This may occur due to database connectivity issues,
                  insufficient metric data, or calculation errors.
                $ref: '#/components/schemas/error'
...
  /api/salesforce/summary/{systemId}:
    get:
      operationId: getSalesforceSummary
      tags:
        - salesforce
      description: Get energy metrics summary for a system from Salesforce
      parameters:
        - schema:
            type: string
          in: path
          name: systemId
          required: true
          description: Unique identifier of the solar system
        - schema:
            type: string
            format: date
          in: query
          name: date
          required: false
          description: Reference date (YYYY-MM-DD format) for the summary
      responses:
        '200':
          description: Successfully retrieved energy metrics summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/summary'
  /api/salesforce/forecast/{systemId}:
    get:
      operationId: getSalesforceForecast
      tags:
        - salesforce
      description: Get energy forecast for a system from Salesforce
      parameters:
        - schema:
            type: string
          in: path
          name: systemId
          required: true
          description: Unique identifier of the solar system
      responses:
        '200':
          description: Successfully retrieved energy forecast
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/forecast'
...